Move local output to own package, refactor, and update check.go
